requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
python-dotenv
mysql-connector-python
pymysql
# -e .